move nearest_point util to track.utils, refactor type hint with optional instead of Union[type, None]
